requirements-geojson2kmlGDrive.txt
functions-framework==3.*
Flask==3.*
google-cloud-storage==2.*
shapely==2.*
google-auth==2.*
google-auth-oauthlib==1.*
google-auth-httplib2==0.*
google-api-python-client==2.*
simplekml==1.*